backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
fastapi==0.115.13
google-auth==2.40.3
google-genai==1.21.1
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
msgpack==1.1.1
numpy==2.3.1
packaging==24.2
pinecone==7.2.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
playwright==1.52.0
portalocker==2.10.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==310
qdrant-client==1.14.3
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
vecx==0.33.1
websockets==15.0.1
